Create musicHelper.py
This program is used to help the user understand where all the notes, as well as what the notes are, for each scale. It tries to give a quick and easy breakdown where the user enters in either "note", where the user then gives the letter of the scale, which can be followed with either flat ("b") or sharp ("#"), the user could also enter "accidental" where the user then has to enter "#" for sharp, "b" for flat, or "0" for no sharps/flats and it will return what scale it is based off of that information.
Once the user knows the scale, it uses the turtle library to draw out a two octave scale, starting in the bass clef and then moving up to the treble clef, as well as showing where the accidentals are.