Create randomTiles.py
This program is where you can put in a set amount width for each square and it will draw out a 400x400 area, with different lengths for each square in the area. It is a fun program to mess around with and then when the squares are done, you can use the get_column_sizes function to get the lengths of each square in the column.