Create passwordGenerator.py
This program is used to generate a password of whatever character length that the user wants, (15 is default character length, but the program will always force it to be equal to or greater than 4) and then the program uses a random number generator to assign a random number, a random character, and a random symbol in 3 different indexes within the password. After this initial stage is done, the program will continue to fill the password with symbols and letters until the password is complete where the function will then print the password to the terminal.