Refactor Game class and add Player management
Refactored Game class to initialize with board and dice instances, added methods for creating the board, getting dice total, and managing player turns. Introduced Player class to manage player state.

Next step:
I will start working on the subclasses within the Space Class that will decide what exactly each type of space does. Examples include:
(Even/Odd) Number - The player must roll either an even number or odd number to move from this space (ex. If the space is set to even, the player must roll either a 2, 4, 6, 8, 10, or 12 to escape this space)
Loop - The player enters a loop and must roll a specific modulo number to escape the loop (ex. The loop is 4 spaces long -> they must roll a 4, 8, 12, 16, 20, etc. to escape the loop)
Skip - If the player lands on this space they get to skip to an automatically set later space on the board (ex. The Skip Start is on space 15 and the skip end is on space 20, if the user lands on 15, they automatically start their next turn on space 20)
Exit - The player must roll an exact number to escape from this space (ex. If the chosen number is 5, the user must roll no more, no less than a 5 to escape and continue playing)